PeerGuardian autoconf and p2pool. USB builder script

+++ guest.sh
@@ -1,7 +1,7 @@
 mount -t proc none /proc
 mount -t sysfs none /sys
 export HOME=/root
-
+export PACKAGES="/tmp/packages"
 
 PURGE_UNNEEDED=true
 ## Purge unnecessary packages
@@ -11,12 +11,9 @@ if $PURGE_UNNEEDED ; then
   sudo apt-get purge $PURGE --assume-yes 
 fi
 
-export PACKAGES="/tmp/packages"
 ADD_NEW=true
 ## Install new packages
-if $ADD_NEW ; then
-  if $UPDATE_APT ; then
-    echo "Adding universe apt repo..."
+if $ADD_NEW ; then if $UPDATE_APT ; then echo "Adding universe apt repo..."
     sudo add-apt-repository "deb http://archive.ubuntu.com/ubuntu $(lsb_release -sc) universe multiverse" 
     sudo apt-get update 
   fi
@@ -32,16 +29,19 @@ if $ADD_NEW ; then
 fi
 
 ## Clean up
-echo "Cleaning up.."
-echo "apt-get clean.."
-sudo apt-get clean 
-echo "apt-get autoremove"
-sudo apt-get autoremove 
-echo "Removing tmp files"
-sudo rm -rf /tmp/*
-sudo rm -f /etc/hosts /etc/resolv.conf
-sudo rm -rf /*old
-sudo rm -rf /*_1
+CLEAN=true
+if $CLEAN; then
+  echo "Cleaning up.."
+  echo "apt-get clean.."
+  sudo apt-get clean 
+  echo "apt-get autoremove"
+  sudo apt-get autoremove 
+  echo "Removing tmp files"
+  sudo rm -rf /tmp/*
+  sudo rm -f /etc/hosts /etc/resolv.conf
+  sudo rm -rf /*old
+  sudo rm -rf /*_1
+fi
 
 ## Unmount proc, sys
 sudo umount /proc || sudo umount -lf /proc
